refactor(protocol): widen imbuement item ids and findItemsById to u32
Item client ids can exceed 65535 on this dat, so the uint16 parameters on the
imbuement-item selection path silently truncated the high bits. Widen them
(and Map::findItemsById, used by helper dummy/exercise scans) to uint32.

- Game::selectImbuementItem / ProtocolGame::sendSelectImbuementItem: u16 -> u32
- Map::findItemsById: clientId u16 -> u32

Target: OTC

+++ src/client/game.cpp
@@ -1561,7 +1561,7 @@ void Game::closeImbuingWindow()
     m_protocolGame->sendCloseImbuingWindow();
 }
 
-void Game::selectImbuementItem(uint16_t itemId, const Position& position, uint8_t stackPos)
+void Game::selectImbuementItem(uint32_t itemId, const Position& position, uint8_t stackPos)
 {
     if (!canPerformGameAction())
         return;
