+++ src/client/game.h
@@ -330,7 +330,7 @@ public:
     void applyImbuement(uint8_t slot, uint32_t imbuementId, bool protectionCharm);
     void clearImbuement(uint8_t slot);
     void closeImbuingWindow();
-    void selectImbuementItem(uint16_t itemId, const Position& position, uint8_t stackPos);
+    void selectImbuementItem(uint32_t itemId, const Position& position, uint8_t stackPos);
     void selectImbuementScroll();
     void imbuementDurations(const bool isOpen = false);
