+++ src/client/protocolgame.h
@@ -141,7 +141,7 @@ public:
     void sendApplyImbuement(uint8_t slot, uint32_t imbuementId, bool protectionCharm);
     void sendClearImbuement(uint8_t slot);
     void sendCloseImbuingWindow();
-    void sendSelectImbuementItem(uint16_t itemId, const Position& position, uint8_t stackPos);
+    void sendSelectImbuementItem(uint32_t itemId, const Position& position, uint8_t stackPos);
     void sendSelectImbuementScroll();
     void sendImbuementDurations(bool isOpen = false);
     void sendAddVip(const std::string& name);
